+++ firmware/hdl/packages/gt_mp7_core_pkg.vhd
@@ -15,11 +15,12 @@
 -- v1.15.2: Added tcl script for "manualy" bit file generation (after timing errors)
 --
 -- frame: v1.2.4 (see frame.vhd)
--- gtl: v1.17.1 (see gtl_module_tpl.vhd)
+-- gtl: v1.17.2 (see gtl_module_tpl.vhd)
 -- fdl: v1.3.6 (see fdl_module.vhd)
 
 -- gtl history:
--- HB 2022-01-20: v3.0.0: Module for "anomaly detection trigger (ADT)" test.
+-- HB 2022-02-17: v3.0.0: Module for "anomaly detection trigger (ADT)" test.
+-- HB 2022-02-16: v1.17.2: Bug fixed in comb_conditions.vhd.
 -- HB 2021-12-09: v1.17.1: Updated logic for jet DISP cut.
 -- HB 2021-10-23: v1.17.0: Implemented logic for jet DISP (displaced) cut [DISP = bit 27 of jet data] in calo_comparators.vhd. Bug fixed in correlation_cuts_wrapper.vhd.
 -- HB 2021-10-27: v1.16.2: Bug fix in matrix_corr_cond.vhd.